Fix int32 overflow in substruct pair indexing
For batches large enough that numTargets * numQueries exceeds INT32_MAX
(e.g. 10M targets x 480 queries = 4.8B pairs), the flat pair index
computed as `targetIdx * numQueries + queryIdx` wrapped to a negative
int. That index was then sign-extended to size_t when used to subscript
HasSubstructMatchResults::hasMatch and the counts vector, causing
out-of-bounds writes and segfaults in hasSubstructMatch and
countSubstructMatches. The same overflow corrupted unordered_map keys
in SubstructSearchResults::matches.

Widen pairIndex computations and the map key type to int64_t. Cast to
size_t at indexing sites. Kernel-side pair indices are unchanged because
they operate in batch-local target coordinates bounded by batchSize.

+++ src/substruct/substruct_search.cu
@@ -1165,7 +1165,7 @@ void hasSubstructMatch(const std::vector<const RDKit::ROMol*>& targets,
 
   for (auto& [pairIdx, matches] : matchResults.matches) {
     if (!matches.empty()) {
-      results.hasMatch[pairIdx] = 1;
+      results.hasMatch[static_cast<size_t>(pairIdx)] = 1;
     }
   }
 }
